app/requirements.txt
flask
prometheus-flask-exporter
flake8
pytest
flask-sqlalchemy
pymysql
cryptography
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-sqlalchemy
